Move DebugCart test to its own file
