+++ use-shopping-cart/react/test/testHelpers.js
@@ -1,0 +1,24 @@
+import * as React from 'react'
+import { CartProvider } from '../index'
+
+export const createWrapper = (overrides = {}) => ({ children }) => (
+  <CartProvider
+    mode="payment"
+    cartMode="client-only"
+    successUrl="https://egghead.io/success"
+    cancelUrl="https://egghead.io/cancel"
+    stripe={null}
+    currency="USD"
+    {...overrides}
+  >
+    {children}
+  </CartProvider>
+)
+
+export const expectedInitialCartState = {
+  cartDetails: {},
+  totalPrice: 0,
+  formattedTotalPrice: '$0.00',
+  cartCount: 0,
+  shouldDisplayCart: false
+}
